Squashed 'airdcpp-webapi/' changes from eae4ca3..dbdfa68
dbdfa68 Remove support for download path auto selection
8bf7585 Refactor priorities, use pointer reference for auto search items in directory downloads
c302816 Update JSON

git-subtree-dir: airdcpp-webapi
git-subtree-split: dbdfa6829d10f6ccc787588924c31b684894d87c

+++ api/QueueFileUtils.cpp
@@ -87,7 +87,7 @@ namespace webserver {
 		switch (aPropertyName) {
 		case PROP_SIZE: return (double)aItem->getSize();
 		case PROP_BYTES_DOWNLOADED: return (double)QueueManager::getInstance()->getDownloadedBytes(aItem);
-		case PROP_PRIORITY: return aItem->getPriority();
+		case PROP_PRIORITY: return (double)aItem->getPriority();
 		case PROP_TIME_ADDED: return (double)aItem->getTimeAdded();
 		case PROP_TIME_FINISHED: return (double)aItem->getTimeFinished();
 		case PROP_SPEED: return (double)QueueManager::getInstance()->getAverageSpeed(aItem);
